Use float division

+++ src/xyz/derkades/ssx_connector/Main.java
@@ -137,7 +137,7 @@ public class Main extends JavaPlugin {
 
 		final int addresses = getConfig().getStringList("addresses").size();
 		final int interval = getConfig().getInt("send-interval");
-		final int taskIntervalTicks = (interval * 20) / addresses;
+		final int taskIntervalTicks = (int) ((interval * 20f) / addresses);
 	
 		this.pingTask = Bukkit.getScheduler().runTaskTimerAsynchronously(this, new PlaceholderSender(), 40, taskIntervalTicks);
 	}
